Restructured the project to follow the hexagonal architecture
