+++ src/main/java/com/quizapp/quiz/adapter/web/dto/ResponseQuestionDTO.java
@@ -1,4 +1,4 @@
-package com.quizapp.quiz.DTOs;
+package com.quizapp.quiz.adapter.web.dto;
 
 import lombok.AllArgsConstructor;
 import lombok.Builder;
